now change the endpoint to accept the data as tolp instead of just id
